# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Digital Inspection App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /maintenance/visits:
    get:
      operationId: listMaintenanceVisits
      tags: [maintenance]
      responses:
        "200":
          description: Visits
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MaintenanceVisit" }
    post:
      operationId: createMaintenanceVisit
      tags: [maintenance]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMaintenanceVisitInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceVisit" }
  /maintenance/visits/{id}:
    get:
      operationId: getMaintenanceVisit
      tags: [maintenance]
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Visit detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceVisitDetail" }
  /maintenance/visits/{id}/venues:
    post:
      operationId: addVenueInspectionVisit
      tags: [maintenance]
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddVenueVisitInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VenueInspectionVisit" }
  /maintenance/repairs:
    get:
      operationId: listRepairs
      tags: [maintenance]
      responses:
        "200":
          description: Repairs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RepairEvent" }
    post:
      operationId: createRepair
      tags: [maintenance]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRepairInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RepairEvent" }
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    MaintenanceVisit:
      type: object
      properties:
        id: { type: integer }
        maintainerId: { type: integer }
        maintainerName: { type: string }
        depotId: { type: integer }
        depotName: { type: string }
        visitDate: { type: string, format: date-time }
        notes:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
      required: [id, maintainerId, maintainerName, depotId, depotName, visitDate, notes, createdAt]
    CreateMaintenanceVisitInput:
      type: object
      properties:
        depotId: { type: integer }
        visitDate:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
      required: [depotId]
    VenueInspectionVisit:
      type: object
      properties:
        id: { type: integer }
        visitId: { type: integer }
        venueId: { type: integer }
        venueName: { type: string }
        venueCode: { type: string }
        visitedAt: { type: string, format: date-time }
        notes:
          type: ["string", "null"]
      required: [id, visitId, venueId, venueName, venueCode, visitedAt, notes]
    AddVenueVisitInput:
      type: object
      properties:
        venueId: { type: integer }
        visitedAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
      required: [venueId]
    RepairEvent:
      type: object
      properties:
        id: { type: integer }
        visitId:
          type: ["integer", "null"]
        venueId: { type: integer }
        venueName: { type: string }
        venueCode:
          type: ["string", "null"]
        assetId:
          type: ["integer", "null"]
        assetLabel:
          type: ["string", "null"]
        action: { type: string, enum: [repair, replace] }
        description:
          type: ["string", "null"]
        partsCostCents: { type: integer }
        labourCostCents: { type: integer }
        clientChargeCents: { type: integer }
        occurredAt: { type: string, format: date-time }
        createdBy:
          type: ["integer", "null"]
        createdByName:
          type: ["string", "null"]
      required: [id, venueId, venueName, action, partsCostCents, labourCostCents, clientChargeCents, occurredAt]
    RepairConsumptionInput:
      type: object
      properties:
        skuId: { type: integer }
        quantity: { type: integer }
      required: [skuId, quantity]
    CreateRepairInput:
      type: object
      properties:
        visitId:
          type: ["integer", "null"]
        venueId: { type: integer }
        assetId:
          type: ["integer", "null"]
        action: { type: string, enum: [repair, replace] }
        description:
          type: ["string", "null"]
        partsCostCents:
          type: ["integer", "null"]
        labourCostCents:
          type: ["integer", "null"]
        clientChargeCents:
          type: ["integer", "null"]
        occurredAt:
          type: ["string", "null"]
          format: date-time
        consumption:
          type: array
          items: { $ref: "#/components/schemas/RepairConsumptionInput" }
      required: [venueId, action]
    MaintenanceVisitDetail:
      type: object
      properties:
        id: { type: integer }
        maintainerId: { type: integer }
        maintainerName: { type: string }
        depotId: { type: integer }
        depotName: { type: string }
        visitDate: { type: string, format: date-time }
        notes:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
        venueVisits:
          type: array
          items: { $ref: "#/components/schemas/VenueInspectionVisit" }
        repairs:
          type: array
          items: { $ref: "#/components/schemas/RepairEvent" }
      required: [id, maintainerId, maintainerName, depotId, depotName, visitDate, notes, createdAt, venueVisits, repairs]