# lib/api-zod/src/generated/types/addVenueVisitInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Inspection App API
 * OpenAPI spec version: 0.1.0
 */

export interface AddVenueVisitInput {
  venueId: number;
  /** @nullable */
  visitedAt?: Date | null;
  /** @nullable */
  notes?: string | null;
}